+++ src/constants.ts
@@ -1,5 +1,6 @@
 export const $signal = Symbol.for('signal');
 export const $effect = Symbol.for('effect');
+
 export const $destroySignal = Symbol.for('destroySignal');
 export const $createEffect = Symbol.for('createEffect');
 export const $destroyEffect = Symbol.for('destroyEffect');
@@ -8,3 +9,5 @@ export const VALUE = 'value';
 export const MUTE = 'mute';
 export const UNMUTE = 'unmute';
 export const DESTROY = 'destroy';
+
+export const RECALL = Symbol.for('recall');