+++ src/global-queues.ts
@@ -3,3 +3,4 @@ import {eventize} from '@spearwolf/eventize';
 export const globalSignalQueue = eventize();
 export const globalDestroySignalQueue = eventize();
 export const globalEffectQueue = eventize();
+export const globalEffectCalledQueue = eventize();